requirements.txt
pipecat-ai[google,sarvam,webrtc,silero]
python-dotenv
loguru
fastapi
uvicorn[standard]
